refactor: clean up $any
