+++ packages/type-plus/src/any/$any.spec.ts
