+++ packages/type-plus/src/type_plus/branch/$select.ts
@@ -1,5 +1,5 @@
 import type { $SpecialType } from '../$special_type.js'
-import type { $Any } from '../../any/any.js'
+import type { $Any } from '../../any/$any.js'
 import type { $Never } from '../../never/never.js'
 import type { $Unknown } from '../../unknown/unknown.js'
 import type { $DistributiveDefault, $DistributiveOptions } from './$distributive.js'
